+++ packages/common/src/index.ts
@@ -1,3 +1,4 @@
 export * from './base-adapter.interface';
 export * from './database-exception.filter';
 export * from './http-error.handler';
+export * from './logger';